\name{AKDE}
\alias{AKDE}
\title{ Adaptive Kernel Density Estimation}
\description{
    It's an adaptive kernel density estimation function. 
}
\usage{
AKDE(tdat, edat)
}
\arguments{
  \item{tdat}{ Numeric. The training data set, matrix.}
  \item{edat}{ Numeric.  Estimating data, vector.}
}
 
\value{A numeric.}
\author{ Yi Ya, Yarong Yang, and Nader Ebrahimi }
\references{Yi Ya, Yarong Yang, Nader Ebrahimi, Yoram Rubin, and Jacob Zhang. SKNN: A Super K-Nearest Neighbor Classification Algorithm. ongoing


Yarong Yang and Nader Ebrahimi (2010). Two Mutual Information Estimators based on Kernel Density Estimation. part of Yarong Yang's Ph.D. dissertation
}



