\name{Dist}
\alias{Dist}
\title{ Finding the distance between two observations.}
\description{
    It's a function of finding the distance between two observations.
}  
\usage{
Dist(x,y)
}
\arguments{
  \item{x}{ Numeric. A vector denoting an observation.}
  \item{y}{ Numeric. A vector denoting an observation.}
}
\value{A list.}
\author{Yi Ya and Yarong Yang }


