\name{SKNN-package}
\alias{SKNN-package}
\alias{SKNN}
\docType{package}
\title{
\packageTitle{SKNN}
}
\description{
\packageDescription{SKNN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SKNN}
\packageIndices{SKNN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SKNN}

Maintainer: \packageMaintainer{SKNN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
